test code to tests sub proj
