+++ modules/tests/shared/src/test/scala/io/circe/JsonSuite.scala
@@ -85,6 +85,55 @@ class JsonSuite extends CirceSuite with FloatJsonTests {
     assert(merged.asObject.map(_.toList) === Some(fields.reverse))
   }
 
+  "deepDropNullValues" should "remove null value for JsonObject" in {
+    val actual = Json
+      .fromFields(
+        List(
+          "a" -> Json.Null,
+          "b" -> Json.fromString("c"),
+          "d" -> Json.fromInt(1),
+          "e" -> Json.True
+        )
+      )
+      .deepDropNullValues
+
+    assert(
+      actual == Json.fromFields(
+        List(
+          "b" -> Json.fromString("c"),
+          "d" -> Json.fromInt(1),
+          "e" -> Json.True
+        )
+      )
+    )
+  }
+
+  "deepDropNullValues" should "remove null value for JsonArray" in {
+    val actual = Json.fromValues(List(Json.Null, Json.fromString("a"))).deepDropNullValues
+
+    assert(actual == Json.fromValues(List(Json.fromString("a"))))
+  }
+
+  "deepDropNullValues" should "remove null value for nested object" in {
+    val actual = Json
+      .fromFields(
+        List(
+          "child1" -> Json.fromFields(List("a" -> Json.Null, "b" -> Json.fromString("c"))),
+          "child2" -> Json.fromValues(List(Json.Null, Json.fromString("a")))
+        )
+      )
+      .deepDropNullValues
+
+    assert(
+      actual == Json.fromFields(
+        List(
+          "child1" -> Json.fromFields(List("b" -> Json.fromString("c"))),
+          "child2" -> Json.fromValues(List(Json.fromString("a")))
+        )
+      )
+    )
+  }
+
   val key = "x"
   val value1 = "fizz"
   val value2 = "foobar"
